+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # GiottoClass 0.3.3
 
+## enhancements
+- `missing` method for `affine()` instantiates an `affine2d` object
 
 # GiottoClass 0.3.2 (2024/07/26)